Rename BookViewModel Class to BookEditViewModel and place it in /Models/BookViewModels Folder.
