+++ src/Open-School-Library/Views/Books/Edit.cshtml
@@ -1,4 +1,4 @@
-@model Open_School_Library.Models.BookViewModel
+@model Open_School_Library.Models.BookViewModels.BookEditViewModel
 
 @{
     ViewData["Title"] = "Edit";
